feat: set up subcommands and restructure for it
