+++ cmd.sh
@@ -1,2 +1,2 @@
-cargo run -- --red-path ./assets/case-02/cutoff_map.png --green-path ./assets/case-02/tissue_argmax.png --blue-path ./assets/case-02/score_image.png --blue-mode "heatmap" --red-bbox 4526 4526 31776 34814  --green-bbox 4539 4539 31774 34800 --blue-bbox  4526 4526 31776 34814 --source-dimensions 37028 35637 --out ./output_02.png
+cargo run -- bitmask-mode --red-path ./assets/case-02/cutoff_map.png --green-path ./assets/case-02/tissue_argmax.png --blue-path ./assets/case-02/score_image.png --blue-mode "heatmap" --red-bbox 4526 4526 31776 34814  --green-bbox 4539 4539 31774 34800 --blue-bbox  4526 4526 31776 34814 --source-dimensions 37028 35637 --out ./output_02.png
