+++ src/dzi_split_mode.rs
@@ -1,0 +1,6 @@
+use crate::app;
+
+
+pub fn do_dzi_split_mode(cli: app::DZISplitModeArgs) {
+    println!("Doing DZI split mode...");
+}
